+++ src/char.diff.gower.c
@@ -16,60 +16,11 @@
 #include "dispRity.h"
 
 // #################
-// Character difference (Gower) logic
+// Character difference logic
 // #################
 
-// Convert character into number
-double character_to_numeric(char c)
-{
-    double n = -1;
-    static const char * const alphabet = "ABCDEFGHIJKLMNOPQRSTUVWXYZ";
-    char *p = strchr(alphabet, (unsigned char)c);
-
-    if (p) {
-        n = p - alphabet;
-    }
-
-    return n;
-}
-
-// Normalise a single numeric character
-void Normalise_single_character(double *vector, int count) {
-    int element, i, j, k;
-    char vector_char[count], element_char;
-    char alphabet[] = { 'A','B','C','D','E','F','G','H','I','J','K','L','M','N','O','P','Q','R','S','T','U','V','W','X','Y','Z','\0'};
-
-    // Getting each unique element of the vector and translating them into standardised characters
-    element = 0;
-
-    for (i = 0; i < count; ++i) {
-        for (j = 0; j < i; ++j) {
-            if (vector[i] == vector[j])
-               break;
-        }
-
-        if (i == j) {
-            // If encountering unique characters, attribute the first letter of the alphabet and so on
-            element_char = alphabet[element];
-            for(k = 0; k < count; k++) {
-                if(vector[k] == vector[i]) {
-                    vector_char[k] = element_char;
-                }
-            }
-            element ++;
-        }
-    }
-
-    // Transforming the character back to numeric (double)
-    for(k = 0; k < count ; k++) {
-        vector[k] = character_to_numeric(vector_char[k]);
-    }
-
-}
-
-
 // Calculating the Gower character distance
-static double R_Gower(double *x, int nr, int nc, int i1, int i2)
+static double R_Gower(double *x, int nr, int nc, int i1, int i2, int translate)
 {
     double diff, dist, vector1[nc], vector2[nc];
     int count, i, k;
@@ -94,15 +45,14 @@ static double R_Gower(double *x, int nr, int nc, int i1, int i2)
     }
 
     // Normalising the characters
-    Normalise_single_character(vector1, count);
-    Normalise_single_character(vector2, count);
+    if(translate) {
+        Normalise_single_character(vector1, count);
+        Normalise_single_character(vector2, count);
+    }
 
+    // Absolute character difference
     for(k = 0 ; k < count ; k++) {
          diff = fabs(vector1[k] - vector2[k]);
-        // Normalise the difference (Fitch like)
-        if(diff > 1) {
-            diff = 1;   
-        }
         // Increment the differences
         if(!ISNAN(diff)) {
             dist += diff;
@@ -112,27 +62,27 @@ static double R_Gower(double *x, int nr, int nc, int i1, int i2)
     if(count == 0) {
         return NA_REAL;
     } else {
-        dist = dist/(count - 1);
+        if(translate) {
+            dist = dist/(count - 1);
+        } else {
+            dist = dist/count;
+        }
         return dist;
     }
 }
 
-// Allowed methods
-enum { GOWER=1};
-/* == 1,2,..., defined by order in the R function dist */
-
-
 // R_distance function (R::dist())
-void R_distance(double *x, int *nr, int *nc, double *d, int *diag, int *method)
+void R_distanceGower(double *x, int *nr, int *nc, double *d, int *diag, int *method, int *translate)
 {
     int dc, i, j;
     size_t  ij;  /* can exceed 2^31 - 1 */
-    double (*distfun)(double*, int, int, int, int) = NULL;
+    double (*distfun)(double*, int, int, int, int, int) = NULL;
 
     //Open MPI
 #ifdef _OPENMP
     int nthreads;
 #endif
+
     distfun = R_Gower;
 
     dc = (*diag) ? 0 : 1; /* diag=1:  we do the diagonal */
@@ -148,7 +98,7 @@ void R_distance(double *x, int *nr, int *nc, double *d, int *diag, int *method)
     ij = 0;
     for(j = 0 ; j <= *nr ; j++)
         for(i = j+dc ; i < *nr ; i++)
-        d[ij++] = distfun(x, *nr, *nc, i, j);
+        d[ij++] = distfun(x, *nr, *nc, i, j, *translate);
     }
     else
     /* This produces uneven thread workloads since the outer loop
@@ -161,23 +111,23 @@ void R_distance(double *x, int *nr, int *nc, double *d, int *diag, int *method)
     for(j = 0 ; j <= *nr ; j++) {
         ij = j * (*nr - dc) + j - ((1 + j) * j) / 2;
         for(i = j+dc ; i < *nr ; i++)
-        d[ij++] = distfun(x, *nr, *nc, i, j);
+        d[ij++] = distfun(x, *nr, *nc, i, j, *translate);
     }
 #else
     ij = 0;
     for(j = 0 ; j <= *nr ; j++)
     for(i = j+dc ; i < *nr ; i++)
-        d[ij++] = distfun(x, *nr, *nc, i, j);
+        d[ij++] = distfun(x, *nr, *nc, i, j, *translate);
 #endif
 }
 
 
 // R/C interface (former Diff)
-SEXP C_char_diff(SEXP x, SEXP smethod, SEXP attrs)
+SEXP C_diff_gower(SEXP x, SEXP smethod, SEXP stranslate, SEXP attrs)
 {
     // Define the variable
     SEXP result;
-    int nr = nrows(x), nc = ncols(x), method = asInteger(smethod);
+    int nr = nrows(x), nc = ncols(x), method = asInteger(smethod), translate = asInteger(stranslate);
     int diag = 0;
     R_xlen_t N;
     N = (R_xlen_t)nr * (nr-1)/2; /* avoid int overflow for N ~ 50,000 */
@@ -186,7 +136,7 @@ SEXP C_char_diff(SEXP x, SEXP smethod, SEXP attrs)
     PROTECT(x);
     
     // Calculate the distance matrix
-    R_distance(REAL(x), &nr, &nc, REAL(result), &diag, &method);
+    R_distanceGower(REAL(x), &nr, &nc, REAL(result), &diag, &method, &translate);
     
     // Wrap up the results
     SEXP names = PROTECT(getAttrib(attrs, R_NamesSymbol)); // Row/column names attributes